refactor: add ArtworkFilterOptionsHeader component
